Factor out common compilation helpers
